+++ DB/DBB/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbb/persistence/db/basic/DbT2003KibetsuDac.java
@@ -6,21 +6,20 @@ package jp.co.ndensan.reams.db.dbb.persistence.db.basic;
 
 import java.util.List;
 import static java.util.Objects.requireNonNull;
-import jp.co.ndensan.reams.db.dbb.entity.db.basic.fuka.DbT2003Kibetsu;
-import static jp.co.ndensan.reams.db.dbb.entity.db.basic.fuka.DbT2003Kibetsu.choshuHouhou;
-import static jp.co.ndensan.reams.db.dbb.entity.db.basic.fuka.DbT2003Kibetsu.choteiNendo;
-import static jp.co.ndensan.reams.db.dbb.entity.db.basic.fuka.DbT2003Kibetsu.fukaNendo;
-import static jp.co.ndensan.reams.db.dbb.entity.db.basic.fuka.DbT2003Kibetsu.ki;
-import static jp.co.ndensan.reams.db.dbb.entity.db.basic.fuka.DbT2003Kibetsu.rirekiNo;
-import static jp.co.ndensan.reams.db.dbb.entity.db.basic.fuka.DbT2003Kibetsu.tsuchishoNo;
-import jp.co.ndensan.reams.db.dbb.entity.db.basic.fuka.DbT2003KibetsuEntity;
+import jp.co.ndensan.reams.db.dbb.entity.db.basic.DbT2003Kibetsu;
+import static jp.co.ndensan.reams.db.dbb.entity.db.basic.DbT2003Kibetsu.choshuHouhou;
+import static jp.co.ndensan.reams.db.dbb.entity.db.basic.DbT2003Kibetsu.choteiNendo;
+import static jp.co.ndensan.reams.db.dbb.entity.db.basic.DbT2003Kibetsu.fukaNendo;
+import static jp.co.ndensan.reams.db.dbb.entity.db.basic.DbT2003Kibetsu.ki;
+import static jp.co.ndensan.reams.db.dbb.entity.db.basic.DbT2003Kibetsu.rirekiNo;
+import static jp.co.ndensan.reams.db.dbb.entity.db.basic.DbT2003Kibetsu.tsuchishoNo;
+import jp.co.ndensan.reams.db.dbb.entity.db.basic.DbT2003KibetsuEntity;
 import jp.co.ndensan.reams.db.dbx.definition.core.valueobject.domain.TsuchishoNo;
 import jp.co.ndensan.reams.db.dbz.persistence.db.basic.ISaveable;
 import jp.co.ndensan.reams.ur.urz.definition.message.UrSystemErrorMessages;
 import jp.co.ndensan.reams.uz.uza.core.mybatis.SqlSession;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleYear;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import jp.co.ndensan.reams.uz.uza.util.db.DbAccessorNormalType;
 import static jp.co.ndensan.reams.uz.uza.util.db.Restrictions.and;
 import static jp.co.ndensan.reams.uz.uza.util.db.Restrictions.eq;
@@ -53,7 +52,7 @@ public class DbT2003KibetsuDac implements ISaveable<DbT2003KibetsuEntity> {
             FlexibleYear 調定年度,
             FlexibleYear 賦課年度,
             TsuchishoNo 通知書番号,
-            Decimal 履歴番号,
+            int 履歴番号,
             RString 徴収方法,
             int 期) throws NullPointerException {
         requireNonNull(調定年度, UrSystemErrorMessages.値がnull.getReplacedMessage("調定年度"));
