+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/kyotakukeika/kyotakukeikakutodokede/KyotakuKeikakuTodokedeBuilder.java
@@ -11,7 +11,7 @@ import jp.co.ndensan.reams.db.dbc.business.core.kyotakukeika.kyotakukeikakujigyo
 import jp.co.ndensan.reams.db.dbc.business.core.kyotakukeika.kyotakukeikakujikosakusei.KyotakuKeikakuJikoSakusei;
 import jp.co.ndensan.reams.db.dbc.business.core.kyotakukeika.kyotakukeikakujikosakusei.KyotakuKeikakuJikoSakuseiIdentifier;
 import jp.co.ndensan.reams.db.dbz.business.core.uzclasses.Models;
-import jp.co.ndensan.reams.db.dbz.entity.db.basic.kyotakukeikaku.DbT3005KyotakuKeikakuTodokedeEntity;
+import jp.co.ndensan.reams.db.dbz.entity.db.basic.DbT3005KyotakuKeikakuTodokedeEntity;
 import jp.co.ndensan.reams.ur.urz.definition.message.UrErrorMessages;
 import jp.co.ndensan.reams.ur.urz.definition.message.UrSystemErrorMessages;
 import jp.co.ndensan.reams.uz.uza.biz.AtenaKanaMeisho;
@@ -205,7 +205,7 @@ public class KyotakuKeikakuTodokedeBuilder {
     private boolean hasSameIdentifier(KyotakuKeikakuJigyoshaSakuseiIdentifier 居宅給付計画事業者作成識別子) {
         return (id.get対象年月().equals(居宅給付計画事業者作成識別子.get対象年月())
                 && id.get被保険者番号().equals(居宅給付計画事業者作成識別子.get被保険者番号())
-                && id.get履歴番号().intValue() == 居宅給付計画事業者作成識別子.get履歴番号());
+                && id.get履歴番号() == 居宅給付計画事業者作成識別子.get履歴番号());
     }
 
     /**
