+++ DB/DBZ/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbz/entity/db/basic/DbT1008IryohokenKanyuJokyoEntity.java
@@ -10,7 +10,6 @@ import jp.co.ndensan.reams.uz.uza.biz.ShikibetsuCode;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 import jp.co.ndensan.reams.uz.uza.lang.RDateTime;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import jp.co.ndensan.reams.uz.uza.util.db.DbTableEntityBase;
 import jp.co.ndensan.reams.uz.uza.util.db.IDbAccessable;
 import jp.co.ndensan.reams.uz.uza.util.db.PrimaryKey;
@@ -37,7 +36,7 @@ public class DbT1008IryohokenKanyuJokyoEntity extends DbTableEntityBase<DbT1008I
     @PrimaryKey
     private ShikibetsuCode shikibetsuCode;
     @PrimaryKey
-    private Decimal rirekiNo;
+    private int rirekiNo;
     private LasdecCode shichosonCode;
     private HihokenshaNo hihokenshaNo;
     private RString iryoHokenShubetsuCode;
@@ -115,7 +114,7 @@ public class DbT1008IryohokenKanyuJokyoEntity extends DbTableEntityBase<DbT1008I
      *
      * @return 履歴番号
      */
-    public Decimal getRirekiNo() {
+    public int getRirekiNo() {
         return rirekiNo;
     }
 
@@ -124,7 +123,7 @@ public class DbT1008IryohokenKanyuJokyoEntity extends DbTableEntityBase<DbT1008I
      *
      * @param rirekiNo 履歴番号
      */
-    public void setRirekiNo(Decimal rirekiNo) {
+    public void setRirekiNo(int rirekiNo) {
         this.rirekiNo = rirekiNo;
     }
