+++ DB/DBZ/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbz/entity/db/basic/DbT3008KyotakuKeikakuJikosakuseiMeisaiEntity.java
@@ -40,7 +40,7 @@ public class DbT3008KyotakuKeikakuJikosakuseiMeisaiEntity extends DbTableEntityB
     @PrimaryKey
     private FlexibleYearMonth taishoYM;
     @PrimaryKey
-    private Decimal rirekiNo;
+    private int rirekiNo;
     @PrimaryKey
     private RString kyotakuServiceKubun;
     @PrimaryKey
@@ -147,7 +147,7 @@ public class DbT3008KyotakuKeikakuJikosakuseiMeisaiEntity extends DbTableEntityB
      *
      * @return 履歴番号
      */
-    public Decimal getRirekiNo() {
+    public int getRirekiNo() {
         return rirekiNo;
     }
 
@@ -156,7 +156,7 @@ public class DbT3008KyotakuKeikakuJikosakuseiMeisaiEntity extends DbTableEntityB
      *
      * @param rirekiNo 履歴番号
      */
-    public void setRirekiNo(Decimal rirekiNo) {
+    public void setRirekiNo(int rirekiNo) {
         this.rirekiNo = rirekiNo;
     }
