+++ DB/DBZ/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbz/persistence/db/basic/DbT1006KyokaisoGaitoshaDac.java
@@ -12,7 +12,6 @@ import static jp.co.ndensan.reams.db.dbz.entity.db.basic.DbT1006KyokaisoGaitosha
 import jp.co.ndensan.reams.db.dbz.entity.db.basic.DbT1006KyokaisoGaitoshaEntity;
 import jp.co.ndensan.reams.ur.urz.definition.message.UrSystemErrorMessages;
 import jp.co.ndensan.reams.uz.uza.core.mybatis.SqlSession;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import jp.co.ndensan.reams.uz.uza.util.db.DbAccessorNormalType;
 import static jp.co.ndensan.reams.uz.uza.util.db.Restrictions.and;
 import static jp.co.ndensan.reams.uz.uza.util.db.Restrictions.eq;
@@ -39,7 +38,7 @@ public class DbT1006KyokaisoGaitoshaDac implements ISaveable<DbT1006KyokaisoGait
     @Transaction
     public DbT1006KyokaisoGaitoshaEntity selectByKey(
             HihokenshaNo 被保険者番号,
-            Decimal 履歴番号) throws NullPointerException {
+            int 履歴番号) throws NullPointerException {
         requireNonNull(被保険者番号, UrSystemErrorMessages.値がnull.getReplacedMessage("被保険者番号"));
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
