+++ DB/DBZ/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbz/persistence/db/basic/DbT3008KyotakuKeikakuJikosakuseiMeisaiDac.java
@@ -50,7 +50,7 @@ public class DbT3008KyotakuKeikakuJikosakuseiMeisaiDac implements ISaveable<DbT3
     public DbT3008KyotakuKeikakuJikosakuseiMeisaiEntity selectByKey(
             HihokenshaNo 被保険者番号,
             FlexibleYearMonth 対象年月,
-            Decimal 履歴番号,
+            int 履歴番号,
             RString 居宅サービス区分,
             JigyoshaNo サービス提供事業者番号,
             ServiceShuruiCode サービス種類コード,
